+++ stories/ElementsStories/DropdownMenu.stories.tsx
@@ -1,14 +1,7 @@
-import type { Meta, StoryObj } from "@storybook/react";
-import {
-  Button,
-  DropdownMenu,
-  Radio,
-  Switch,
-  MenuItemType,
-  Chip,
-} from "../../components/elements";
+import { useState } from "react";
+
 import { ArgsTable, Story } from "@storybook/blocks";
-import { setLocale, t } from "../translations/i18n";
+import type { Meta, StoryObj } from "@storybook/react";
 import {
   AlignVerticalDistributeStart,
   AtSign,
@@ -16,9 +9,19 @@ import {
   FolderArchive,
   FolderOpen,
   Home,
-  MedalIcon,
+  MedalIcon
 } from "lucide-react";
-import { useState } from "react";
+
+import {
+  Button,
+  DropdownMenu,
+  Radio,
+  Switch,
+  MenuItemType,
+  Chip,
+  DropdownMenuRadio
+} from "../../components/elements";
+import { setLocale, t } from "../translations/i18n";
 
 const meta = {
   title: "Elements/DropdownMenu",
@@ -31,10 +34,10 @@ const meta = {
           <h1>{"<DropdownMenu/>"}</h1>
           <ArgsTable />
         </>
-      ),
-    },
+      )
+    }
   },
-  tags: ["autodocs"],
+  tags: ["autodocs"]
 } satisfies Meta<typeof DropdownMenu>;
 
 export default meta;
@@ -46,7 +49,7 @@ const items = [
     value: "item1",
     onMiddleClick: (e: any) => console.log("middle clicked ", e),
     icon: <Home className="hawa-icon" />,
-    shortcut: <span>shift + E</span>,
+    shortcut: <span>shift + E</span>
   },
 
   {
@@ -55,14 +58,14 @@ const items = [
     onMiddleClick: (e: any) => console.log("middle clicked ", e),
     icon: <Bell className="hawa-icon" />,
     badged: true,
-    end: <span>shift + E</span>,
+    end: <span>shift + E</span>
   },
   {
     label: "With Badge",
     value: "item_badged",
     onMiddleClick: (e: any) => console.log("middle clicked ", e),
     icon: <Bell className="hawa-icon" />,
-    badged: true,
+    badged: true
   },
   {
     label: "Disabled",
@@ -70,14 +73,14 @@ const items = [
     icon: <FolderOpen className="hawa-icon" />,
     end: <Switch size="sm" />,
     presist: true,
-    disabled: true,
+    disabled: true
   },
   {
     label: "Small Switch",
     value: "item2",
     icon: <FolderOpen className="hawa-icon" />,
     end: <Switch size="sm" />,
-    presist: true,
+    presist: true
   },
 
   {
@@ -97,7 +100,7 @@ const items = [
               >
                 <path d="M6.993 12c0 2.761 2.246 5.007 5.007 5.007s5.007-2.246 5.007-5.007S14.761 6.993 12 6.993 6.993 9.239 6.993 12zM12 8.993c1.658 0 3.007 1.349 3.007 3.007S13.658 15.007 12 15.007 8.993 13.658 8.993 12 10.342 8.993 12 8.993zM10.998 19h2v3h-2zm0-17h2v3h-2zm-9 9h3v2h-3zm17 0h3v2h-3zM4.219 18.363l2.12-2.122 1.415 1.414-2.12 2.122zM16.24 6.344l2.122-2.122 1.414 1.414-2.122 2.122zM6.342 7.759 4.22 5.637l1.415-1.414 2.12 2.122zm13.434 10.605-1.414 1.414-2.122-2.122 1.414-1.414z"></path>
               </svg>
-            ),
+            )
           },
           {
             value: "en",
@@ -111,34 +114,34 @@ const items = [
               >
                 <path d="M9.37 5.51c-.18.64-.27 1.31-.27 1.99 0 4.08 3.32 7.4 7.4 7.4.68 0 1.35-.09 1.99-.27C17.45 17.19 14.93 19 12 19c-3.86 0-7-3.14-7-7 0-2.93 1.81-5.45 4.37-6.49zM12 3c-4.97 0-9 4.03-9 9s4.03 9 9 9 9-4.03 9-9c0-.46-.04-.92-.1-1.36-.98 1.37-2.58 2.26-4.4 2.26-2.98 0-5.4-2.42-5.4-5.4 0-1.81.89-3.42 2.26-4.4-.44-.06-.9-.1-1.36-.1z"></path>
               </svg>
-            ),
-          },
+            )
+          }
         ]}
       />
     ),
-    presist: true,
+    presist: true
   },
   {
     label: "With Switch",
     value: "item2",
     icon: <FolderOpen className="hawa-icon" />,
     end: <Switch />,
-    presist: true,
+    presist: true
   },
   {
-    itemType: "separator",
+    itemType: "separator"
   },
   {
     label: "Label Here",
     value: "item4",
-    itemType: "label",
+    itemType: "label"
   },
   {
     label: "Item 2",
     value: "item2",
     icon: <FolderArchive className="hawa-icon" />,
     end: <Switch />,
-    presist: true,
+    presist: true
   },
   {
     presist: true,
@@ -149,10 +152,10 @@ const items = [
         design="tabs"
         options={[
           { value: "ar", label: "عربي" },
-          { value: "en", label: "English" },
+          { value: "en", label: "English" }
         ]}
       />
-    ),
+    )
   },
   {
     label: "Item 3",
@@ -164,15 +167,15 @@ const items = [
         label: "subitem 1",
         value: "subitem 1",
         onMiddleClick: (e: any) => console.log("test", e),
-        icon: <MedalIcon className="hawa-icon" />,
+        icon: <MedalIcon className="hawa-icon" />
       },
       {
         label: "subitem 2",
         value: "subitem 2",
         onMiddleClick: () => console.log("test"),
-        icon: <AlignVerticalDistributeStart className="hawa-icon" />,
-      },
-    ],
+        icon: <AlignVerticalDistributeStart className="hawa-icon" />
+      }
+    ]
   },
   {
     label: "Item 3",
@@ -182,29 +185,29 @@ const items = [
       {
         label: "subitem 1",
         value: "subitem 1",
-        icon: <MedalIcon className="hawa-icon" />,
+        icon: <MedalIcon className="hawa-icon" />
       },
       {
         label: "subitem 2",
         value: "subitem 2",
-        icon: <AlignVerticalDistributeStart className="hawa-icon" />,
-      },
-    ],
-  },
+        icon: <AlignVerticalDistributeStart className="hawa-icon" />
+      }
+    ]
+  }
 ];
 const items2 = [
   {
     itemType: "custom",
     content: (
-      <div className="hawa-p-2 hawa-flex hawa-flex-row hawa-gap-2 hawa-items-center hawa-justify-center hover:hawa-bg-muted">
+      <div className="hawa-flex hawa-flex-row hawa-items-center hawa-justify-center hawa-gap-2 hawa-p-2 hover:hawa-bg-muted">
         <FolderOpen />
         <div className="hawa-flex hawa-flex-col hawa-justify-center ">
-          <div className="hawa-font-bold hawa-text-lg">Custom Title</div>
+          <div className="hawa-text-lg hawa-font-bold">Custom Title</div>
           <div className="hawa-text-sm">Subtitle of this menu item here</div>
         </div>
       </div>
-    ),
-  },
+    )
+  }
 ];
 
 export const Default: Story = {
@@ -229,7 +232,7 @@ export const Default: Story = {
         </div>
       </div>
     );
-  },
+  }
 };
 export const Direction: Story = {
   render: (args: any, globals: any) => {
@@ -257,7 +260,7 @@ export const Direction: Story = {
         </div>
       </div>
     );
-  },
+  }
 };
 export const Sizes: Story = {
   render: (args: any, globals: any) => {
@@ -270,7 +273,7 @@ export const Sizes: Story = {
 
     return (
       <div className="hawa-flex  hawa-flex-1 hawa-flex-col ">
-        <div className="hawa-flex-row hawa-flex hawa-w-full hawa-items-center hawa-justify-center hawa-gap-6 hawa-p-3">
+        <div className="hawa-flex hawa-w-full hawa-flex-row hawa-items-center hawa-justify-center hawa-gap-6 hawa-p-3">
           <DropdownMenu
             // width="parent"
             trigger={<Button>Default Size</Button>}
@@ -285,7 +288,7 @@ export const Sizes: Story = {
             onItemSelect={handleItemSelect}
           />
         </div>
-        <div className="hawa-flex-row hawa-flex hawa-w-full hawa-items-center hawa-justify-center hawa-gap-6 hawa-p-3">
+        <div className="hawa-flex hawa-w-full hawa-flex-row hawa-items-center hawa-justify-center hawa-gap-6 hawa-p-3">
           <DropdownMenu
             width="parent"
             trigger={<Button>Menu Width Equal to Trigger Button</Button>}
@@ -295,7 +298,7 @@ export const Sizes: Story = {
         </div>
       </div>
     );
-  },
+  }
 };
 export const WithHeader: Story = {
   render: (args: any, globals: any) => {
@@ -308,11 +311,11 @@ export const WithHeader: Story = {
 
     return (
       <div className="hawa-flex  hawa-flex-1 hawa-flex-col ">
-        <div className="hawa-flex-row hawa-flex hawa-w-full hawa-items-center hawa-justify-center hawa-gap-6 hawa-p-3">
+        <div className="hawa-flex hawa-w-full hawa-flex-row hawa-items-center hawa-justify-center hawa-gap-6 hawa-p-3">
           <DropdownMenu
             header={
               <div className="hawa-p-2">
-                <div className="hawa-font-bold hawa-text-lg">
+                <div className="hawa-text-lg hawa-font-bold">
                   Sikka Software
                 </div>
                 <div className="hawa-flex hawa-flex-row hawa-items-center hawa-gap-2">
@@ -329,5 +332,34 @@ export const WithHeader: Story = {
         </div>
       </div>
     );
-  },
+  }
+};
+export const RadioMenu: Story = {
+  render: (args: any, globals: any) => {
+    const locale = globals.globals?.locale === "ar" ? "ar" : "en";
+    setLocale(locale);
+
+    const handleItemSelect = (value: any) => {
+      console.log("Selected item:", value);
+    };
+
+    const [showPopup, setShowPopup] = useState(false);
+    const [inputLang, setInputLang] = useState("en");
+    return (
+      <DropdownMenuRadio
+        label="Input Language"
+        value={inputLang}
+        onValueChange={setInputLang}
+        options={[
+          { label: "Arabic", value: "ar" },
+          { label: "English", value: "en" }
+        ]}
+        trigger={
+          <Button onClick={() => setShowPopup(!showPopup)}>
+            Open Radio Menu
+          </Button>
+        }
+      />
+    );
+  }
 };