+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sikka/hawa
 
+## 0.21.5
+
+### Patch Changes
+
+- Added `DropdownMenuRadio` to a storybook
+- Fixed cursor design in `DropdownMenuRadio`
+
 ## 0.21.4
 
 ### Patch Changes